+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ADDED:
 
 - `AbstractSimpleResourceDocument` to define documents containing information about a single resource without
 the need of a resource transformer
+- `ClientGeneratedIdRequired` exception
 
 CHANGED: